Comeonin behaviour (Comeonin v5.5.1)

Defines a behaviour for higher-level password hashing functions.

Types

opts()

@type opts() :: keyword()

password()

@type password() :: binary()

user_struct()

@type user_struct() :: map() | nil

Callbacks

add_hash(password, opts)

This callback is deprecated. This function will be removed in the next major version..

@callback add_hash(password(), opts()) :: map()

check_pass(user_struct, password, opts)

This callback is deprecated. This function will be removed in the next major version..

@callback check_pass(user_struct(), password(), opts()) ::
  {:ok, map()} | {:error, String.t()}

no_user_verify(opts)

@callback no_user_verify(opts()) :: false

Runs the password hash function, but always returns false.

This function is intended to make it more difficult for any potential attacker to find valid usernames by using timing attacks. This function is only useful if it is used as part of a policy of hiding usernames.