Comeonin.PasswordHash behaviour (Comeonin v5.5.1)

Defines a behaviour for password hashing functions.

Types

opts()

@type opts() :: keyword()

password()

@type password() :: binary()

password_hash()

@type password_hash() :: binary()

Callbacks

hash_pwd_salt(password, opts)

@callback hash_pwd_salt(password(), opts()) :: password_hash()

Generates a random salt and then hashes the password.

verify_pass(password, password_hash)

@callback verify_pass(password(), password_hash()) :: boolean()

Checks the password by comparing it with a stored hash.

Please note that the first argument to verify_pass should be the password, and the second argument should be the password hash.